Add files via upload
This PHP project lets the bot analyze the user’s messages and the replies to the questions are sent accordingly. This web app can be utilized to respond to the queries of the students. You can create a college query bot of sorts thereby making it easier for the students. With artificial intelligence, the chatbot will be able to analyze the questions of the users efficiently and will provide appropriate responses. If the chatbot is incapable to provide suitable responses to the query of the user, then they can give feedback to enhance the accuracy. For this project, you get to implement the knowledge you have of PHP, UI, and AI for this project.